Fix to module loading error handling.
In `PluginManager#_loadModulesFromDirectory()`, a missing module is
supposed to produce a warning message, but other errors are supposed to
produce an error message and force the process to exit. However, if a
module requires a subsidiary module that is missing, this too just
generates a warning, when it should really be an error.

Changed the code to distinguish the different cases, and tweaked the
messages. The test is a kludge, but I couldn't come up with anything
better.