Added documentation and tests for Application start() method behavior.
The promise returned by the `Application` `start()` method resolves when
the application boot sequence completes (`restart()` behaves the same
way).

This came up in writing test cases that needed to wait for the
application to boot. Synchronizing with `start()` is quite a bit
simpler/cleaner than using a `onceAfter('launch')` event handler. Don't
know if it's useful anywhere else.